requirements.txt
apache-airflow==2.7.2
kafka-python
dbt-core
dbt-postgres
psycopg2-binary
